feat: refactoring + added unit tests for lightspeed package
