+++ cmd/lightspeed.go
@@ -1,375 +1,259 @@
-// Generated by Claude Code
-// RAG model deployment functions for krknctl lightspeed
-
+// Package cmd contains the CLI commands for krknctl, including GPU and acceleration utilities.
+// Assisted-by: Claude Sonnet 4
 package cmd
 
 import (
-	"bufio"
-	"bytes"
-	"context"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"os/signal"
-	"strings"
-	"syscall"
-	"time"
+	"github.com/krkn-chaos/krknctl/pkg/lightspeed"
 
 	"github.com/krkn-chaos/krknctl/pkg/config"
-	"github.com/krkn-chaos/krknctl/pkg/gpucheck"
+	"github.com/krkn-chaos/krknctl/pkg/provider/factory"
 	"github.com/krkn-chaos/krknctl/pkg/provider/models"
 	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator"
+	orchestratormodels "github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/models"
+	"github.com/spf13/cobra"
 )
 
-// RAGDeploymentResult holds information about the deployed RAG model
-type RAGDeploymentResult struct {
-	containerID string
-	hostPort    string
+func NewLightspeedCommand() *cobra.Command {
+	var command = &cobra.Command{
+		Use:   "lightspeed",
+		Short: "GPU and acceleration related utilities",
+		Long: `GPU and acceleration related utilities for container runtime environments
+
+Available Commands:
+  check    Check GPU support in container runtime
+  run      Run AI-powered chaos engineering assistance
+
+GPU Auto-Detection:
+  Lightspeed automatically detects your GPU type! No manual flags needed.
+  
+  Supported GPU Types:
+  • Apple Silicon (M1, M2, M3, M4 with Metal)
+  • NVIDIA GPUs (CUDA, GeForce, Quadro, Tesla)
+  
+  Future support planned:
+  • AMD GPUs (Radeon, FirePro, Instinct with ROCm)
+  • Intel GPUs (Arc, Iris, UHD Graphics)
+
+Examples:
+  krknctl lightspeed check
+  krknctl lightspeed run
+  krknctl lightspeed run --no-gpu`,
+	}
+
+	// GPU auto-detection - no manual flags needed anymore!
+
+	// Add no-gpu flag for CPU-only mode
+	command.PersistentFlags().Bool("no-gpu", false, "Use CPU-only mode (no GPU acceleration)")
+
+	return command
 }
 
-// deployRAGModelWithGPUType deploys the RAG model container using the new GPU detection system
-func deployRAGModelWithGPUType(ctx context.Context, gpuType gpucheck.GPUAcceleration, orchestrator scenarioorchestrator.ScenarioOrchestrator, config config.Config, registry *models.RegistryV2, detector *gpucheck.PlatformGPUDetector) (*RAGDeploymentResult, error) {
-	// Get the appropriate lightspeed image for the detected GPU type
-	ragImageURI, err := detector.GetLightspeedImageURI(gpuType)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get lightspeed image URI: %w", err)
-	}
+// GPU auto-detection - no manual flag parsing needed anymore!
 
-	// Generate unique container name using config
-	containerName := fmt.Sprintf("%s-%d", config.RAGContainerPrefix, time.Now().Unix())
+func NewLightspeedCheckCommand(
+	providerFactory *factory.ProviderFactory,
+	scenarioOrchestrator *scenarioorchestrator.ScenarioOrchestrator,
+	config config.Config,
+) *cobra.Command {
+	var command = &cobra.Command{
+		Use:   "check",
+		Short: "Check GPU support in container runtime",
+		Long: `Check whether the container runtime (Podman or Docker) has GPU support available
 
-	// Set up environment variables
-	env := map[string]string{}
+Lightspeed automatically tests all supported GPU types to detect your hardware.
 
-	// Get device mounts from detector
-	devices := detector.GetDeviceMounts(gpuType)
+Examples:
+  krknctl lightspeed check`,
+		Args: cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			// Print container runtime info
+			(*scenarioOrchestrator).PrintContainerRuntime()
 
-	// Set up port mapping using config
-	hostPort := config.RAGServicePort      // Host port (e.g., "8080")
-	containerPort := config.RAGServicePort // Container port (e.g., "8080")
-	portMappings := &map[string]string{
-		hostPort: containerPort, // host port -> container port
-	}
-
-	// Create spinner for pull progress (exactly like run.go)
-	spinner := NewSpinnerWithSuffix("pulling RAG model image...")
-	spinner.Start()
-
-	// Create communication channel for pull progress updates
-	commChan := make(chan *string)
-	go func() {
-		for msg := range commChan {
-			spinner.Suffix = *msg
-		}
-		spinner.Stop()
-	}()
-
-	// Run the RAG container in detached mode
-	containerID, err := orchestrator.Run(ragImageURI, containerName, env, false, nil, &devices,
-		&commChan, ctx, registry, portMappings)
-
-	// The orchestrator closes the channel automatically, so we don't need to close it manually
-	if err != nil {
-		return nil, fmt.Errorf("failed to run RAG container: %w", err)
-	}
-	fmt.Printf("🚀 RAG container started: %s\n", *containerID)
-	fmt.Printf("📡 Port mapping: %s:%s -> container:%s\n", config.RAGHost, hostPort, config.RAGServicePort)
-
-	return &RAGDeploymentResult{
-		containerID: *containerID,
-		hostPort:    hostPort,
-	}, nil
-}
-
-// deployRAGModel deploys the RAG model container in detached mode with proper port mapping
-// DEPRECATED: Use deployRAGModelWithGPUType instead
-func deployRAGModel(ctx context.Context, gpuType string, orchestrator scenarioorchestrator.ScenarioOrchestrator, config config.Config, registry *models.RegistryV2) (*RAGDeploymentResult, error) {
-	// Get RAG model image URI
-	ragImageURI, err := config.GetRAGModelImageURI()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get RAG model image URI: %w", err)
-	}
-
-	// Generate unique container name using config
-	containerName := fmt.Sprintf("%s-%d", config.RAGContainerPrefix, time.Now().Unix())
-
-	// Set up environment variables
-	env := map[string]string{}
-
-	// Set up GPU device mounts (legacy approach)
-	devices := make(map[string]string)
-	switch gpuType {
-	case "nvidia":
-		devices["/dev/nvidia0"] = "/dev/nvidia0"
-		devices["/dev/nvidiactl"] = "/dev/nvidiactl"
-		devices["/dev/nvidia-uvm"] = "/dev/nvidia-uvm"
-	case "apple-silicon":
-		devices["/dev/dri"] = "/dev/dri"
-	}
-
-	// Set up port mapping using config
-	hostPort := config.RAGServicePort      // Host port (e.g., "8080")
-	containerPort := config.RAGServicePort // Container port (e.g., "8080")
-	portMappings := &map[string]string{
-		hostPort: containerPort, // host port -> container port
-	}
-
-	// Create spinner for pull progress (exactly like run.go)
-	spinner := NewSpinnerWithSuffix("pulling RAG model image...")
-	spinner.Start()
-
-	// Create communication channel for pull progress updates
-	commChan := make(chan *string)
-	go func() {
-		for msg := range commChan {
-			spinner.Suffix = *msg
-		}
-		spinner.Stop()
-	}()
-
-	// Run the RAG container in detached mode
-	containerID, err := orchestrator.Run(ragImageURI, containerName, env, false, nil, &devices,
-		&commChan, ctx, registry, portMappings)
-
-	// The orchestrator closes the channel automatically, so we don't need to close it manually
-	if err != nil {
-		return nil, fmt.Errorf("failed to run RAG container: %w", err)
-	}
-	fmt.Printf("🚀 RAG container started: %s\n", *containerID)
-	fmt.Printf("📡 Port mapping: %s:%s -> container:%s\n", config.RAGHost, hostPort, config.RAGServicePort)
-
-	return &RAGDeploymentResult{
-		containerID: *containerID,
-		hostPort:    hostPort,
-	}, nil
-}
-
-
-// HealthResponse represents the health check response
-type HealthResponse struct {
-	Status           string `json:"status"`
-	Service          string `json:"service"`
-	Model            string `json:"model"`
-	DocumentsIndexed int    `json:"documents_indexed"`
-}
-
-// performRAGHealthCheck performs health checking with timeout and cleanup on failure
-func performRAGHealthCheck(containerID string, hostPort string, orchestrator scenarioorchestrator.ScenarioOrchestrator, ctx context.Context, config config.Config) (bool, error) {
-	// Health URL is constructed from trusted config values and validated port
-	healthURL := fmt.Sprintf("http://%s:%s%s", config.RAGHost, hostPort, config.RAGHealthEndpoint)
-	maxRetries := config.RAGHealthMaxRetries
-	retryInterval := time.Duration(config.RAGHealthRetryIntervalSeconds) * time.Second
-
-	fmt.Printf("🩺 Health checking Lightspeed service at %s...\n", healthURL)
-
-	for i := 0; i < maxRetries; i++ {
-		// Check if container is still running
-		containers, err := orchestrator.ListRunningContainers(ctx)
-		if err != nil {
-			fmt.Printf("⚠️  Warning: failed to list containers: %v\n", err)
-		} else {
-			containerFound := false
-			if containers != nil {
-				for _, container := range *containers {
-					if container.ID == containerID {
-						containerFound = true
-						break
-					}
-				}
+			// Check if Docker is being used - Lightspeed only supports Podman
+			if (*scenarioOrchestrator).GetContainerRuntime() == orchestratormodels.Docker {
+				return fmt.Errorf("❌ Lightspeed requires Podman container runtime. Docker is not supported for GPU acceleration")
 			}
-			if !containerFound {
-				return false, fmt.Errorf("container %s is no longer running", containerID)
+
+			// Get container runtime socket
+			socket, err := (*scenarioOrchestrator).GetContainerRuntimeSocket(nil)
+			if err != nil {
+				return fmt.Errorf("failed to get container runtime socket: %w", err)
 			}
-		}
 
-		// Perform health check - URL is safe as it's constructed from config
-		// #nosec G107 - URL constructed from trusted configuration
-		resp, err := http.Get(healthURL)
-		if err == nil && resp.StatusCode == 200 {
-			var health HealthResponse
-			if err := json.NewDecoder(resp.Body).Decode(&health); err == nil {
-				if closeErr := resp.Body.Close(); closeErr != nil {
-					fmt.Printf("⚠️  Warning: failed to close response body: %v\n", closeErr)
-				}
-				if health.Status == "healthy" {
-					fmt.Printf("✅ Service healthy: %s with %d documents indexed\n", health.Model, health.DocumentsIndexed)
-					return true, nil
-				}
+			// Connect to container runtime
+			ctx, err := (*scenarioOrchestrator).Connect(*socket)
+			if err != nil {
+				return fmt.Errorf("failed to connect to container runtime: %w", err)
 			}
-			if closeErr := resp.Body.Close(); closeErr != nil {
-				fmt.Printf("⚠️  Warning: failed to close response body: %v\n", closeErr)
+
+			// Get no-gpu flag
+			noGPU, _ := cmd.Flags().GetBool("no-gpu")
+
+			// Create platform GPU detector
+			detector := lightspeed.NewPlatformGPUDetector(config)
+
+			// Auto-detect GPU acceleration
+			fmt.Println("\n🔍 Detecting GPU acceleration...")
+			gpuType := detector.DetectGPUAcceleration(ctx, noGPU)
+
+			// Get configuration
+			imageURI, _, deviceMounts, err := detector.AutoSelectLightspeedConfig(ctx, noGPU)
+			if err != nil {
+				return fmt.Errorf("failed to get lightspeed configuration: %w", err)
 			}
-		}
 
-		if i < maxRetries-1 {
-			fmt.Printf("⏳ Waiting for service to become ready... (%d/%d)\n", i+1, maxRetries)
-			time.Sleep(retryInterval)
-		}
-	}
+			// Format and print result
+			fmt.Printf("\n✅ GPU acceleration: %s\n", detector.GetGPUDescription(gpuType))
+			fmt.Printf("📦 Container image: %s\n", imageURI)
+			if len(deviceMounts) > 0 {
+				fmt.Printf("🔗 Device mounts: %v\n", deviceMounts)
+			} else {
+				fmt.Printf("🔗 Device mounts: none (CPU-only)\n")
+			}
 
-	// Health check failed, clean up the container
-	timeoutMinutes := (maxRetries * config.RAGHealthRetryIntervalSeconds) / 60
-	fmt.Printf("❌ Health check timed out after %d minutes, cleaning up container...\n", timeoutMinutes)
-	if err := orchestrator.Kill(&containerID, ctx); err != nil {
-		fmt.Printf("⚠️  Warning: failed to clean up container %s: %v\n", containerID, err)
-	}
-
-	return false, fmt.Errorf("service did not become healthy within timeout")
-}
-
-// ChatMessage represents a message in OpenAI format
-type ChatMessage struct {
-	Role    string `json:"role"`
-	Content string `json:"content"`
-}
-
-// QueryRequest represents an OpenAI-compatible chat completion request
-type QueryRequest struct {
-	Model       string        `json:"model"`
-	Messages    []ChatMessage `json:"messages"`
-	Temperature float64       `json:"temperature,omitempty"`
-	MaxTokens   int           `json:"max_tokens,omitempty"`
-	Stream      bool          `json:"stream,omitempty"`
-}
-
-// QueryChoice represents a choice in OpenAI response format
-type QueryChoice struct {
-	Index   int `json:"index"`
-	Message struct {
-		Role    string `json:"role"`
-		Content string `json:"content"`
-	} `json:"message"`
-	FinishReason string `json:"finish_reason"`
-}
-
-// QueryResponse represents an OpenAI-compatible response
-type QueryResponse struct {
-	ID      string        `json:"id"`
-	Object  string        `json:"object"`
-	Created int64         `json:"created"`
-	Model   string        `json:"model"`
-	Choices []QueryChoice `json:"choices"`
-	Usage   struct {
-		PromptTokens     int `json:"prompt_tokens"`
-		CompletionTokens int `json:"completion_tokens"`
-		TotalTokens      int `json:"total_tokens"`
-	} `json:"usage"`
-}
-
-// startInteractivePrompt starts an interactive chat session with the Lightspeed service
-func startInteractivePrompt(containerID string, hostPort string, orchestrator scenarioorchestrator.ScenarioOrchestrator, ctx context.Context, config config.Config) error {
-	scanner := bufio.NewScanner(os.Stdin)
-
-	// Set up signal handling for Ctrl+C
-	signalChan := make(chan os.Signal, 1)
-	signal.Notify(signalChan, os.Interrupt, syscall.SIGTERM)
-
-	// Handle Ctrl+C in a goroutine
-	go func() {
-		<-signalChan
-		fmt.Println("\n\n🛑 Interrupted! Cleaning up Lightspeed service...")
-		if err := orchestrator.Kill(&containerID, ctx); err != nil {
-			fmt.Printf("⚠️  Warning: failed to stop Lightspeed container: %v\n", err)
-		} else {
-			fmt.Println("✅ Lightspeed service stopped successfully")
-		}
-		os.Exit(0)
-	}()
-
-	fmt.Printf("🤖 AI Assistant ready! Ask me about krknctl commands or chaos engineering:\n")
-	fmt.Printf("📍 Service available at: http://%s:%s\n", config.RAGHost, hostPort)
-	fmt.Printf("💡 Try asking: 'How do I run a pod deletion scenario?'\n")
-	fmt.Printf("🚪 Type 'exit', 'quit', or press Ctrl+C to stop.\n\n")
-
-	for {
-		fmt.Print("> ")
-		if !scanner.Scan() {
-			break
-		}
-
-		query := strings.TrimSpace(scanner.Text())
-		if query == "" {
-			continue
-		}
-
-		if strings.ToLower(query) == "exit" || strings.ToLower(query) == "quit" {
-			fmt.Println("👋 Goodbye!")
-			break
-		}
-
-		// Send query to Lightspeed service
-		response, err := queryRAGService(hostPort, query, config)
-		if err != nil {
-			fmt.Printf("❌ Error: %v\n", err)
-			fmt.Printf("💡 Tip: Make sure the service is fully initialized. Large models may take a few minutes.\n")
-			continue
-		}
-
-		// Display response
-		if len(response.Choices) > 0 {
-			fmt.Printf("\n🤖 %s\n", response.Choices[0].Message.Content)
-		} else {
-			fmt.Printf("\n🤖 No response received\n")
-		}
-		fmt.Println()
-	}
-
-	if err := scanner.Err(); err != nil {
-		return fmt.Errorf("error reading input: %w", err)
-	}
-
-	// Clean up: stop the container
-	fmt.Println("\n🧹 Cleaning up Lightspeed service...")
-	if err := orchestrator.Kill(&containerID, ctx); err != nil {
-		fmt.Printf("⚠️  Warning: failed to stop Lightspeed container: %v\n", err)
-	} else {
-		fmt.Println("✅ Lightspeed service stopped successfully")
-	}
-
-	return nil
-}
-
-// queryRAGService sends a query to the Lightspeed service and returns the response
-func queryRAGService(hostPort string, query string, config config.Config) (*QueryResponse, error) {
-	// Query URL is constructed from trusted config values and validated port
-	url := fmt.Sprintf("http://%s:%s%s", config.RAGHost, hostPort, config.RAGQueryEndpoint)
-
-	requestBody := QueryRequest{
-		Model: "llama",
-		Messages: []ChatMessage{
-			{
-				Role:    "user",
-				Content: query,
-			},
+			return nil
 		},
-		Temperature: 0.7,
-		MaxTokens:   512,
-		Stream:      false,
 	}
 
-	jsonData, err := json.Marshal(requestBody)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal request: %w", err)
+	return command
+}
+
+// buildLightspeedRegistryFromFlags builds lightspeed registry configuration from command flags
+func buildLightspeedRegistryFromFlags(cmd *cobra.Command, config config.Config) (*models.RegistryV2, error) {
+	// Get private registry flags from parent command
+	privateRegistry, _ := cmd.Flags().GetString("private-registry")
+	privateRegistryLightspeed, _ := cmd.Flags().GetString("private-registry-lightspeed")
+	privateRegistryUsername, _ := cmd.Flags().GetString("private-registry-username")
+	privateRegistryPassword, _ := cmd.Flags().GetString("private-registry-password")
+	privateRegistryToken, _ := cmd.Flags().GetString("private-registry-token")
+	privateRegistryInsecure, _ := cmd.Flags().GetBool("private-registry-insecure")
+	privateRegistrySkipTLS, _ := cmd.Flags().GetBool("private-registry-skip-tls")
+
+	// If no private registry is specified, return nil (use default public registry)
+	if privateRegistry == "" {
+		return nil, nil
 	}
 
-	// #nosec G107 - URL constructed from trusted configuration
-	resp, err := http.Post(url, "application/json", bytes.NewBuffer(jsonData))
-	if err != nil {
-		return nil, fmt.Errorf("failed to send request: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("service returned error %d: %s", resp.StatusCode, string(body))
+	// Use lightspeed repository override if provided, otherwise use config default
+	lightspeedRepo := privateRegistryLightspeed
+	if lightspeedRepo == "" {
+		lightspeedRepo = config.LightspeedRegistry
 	}
 
-	var response QueryResponse
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		return nil, fmt.Errorf("failed to decode response: %w", err)
+	// Build registry configuration
+	registry := &models.RegistryV2{
+		RegistryURL:        privateRegistry,
+		Username:           &privateRegistryUsername,
+		Password:           &privateRegistryPassword,
+		Token:              &privateRegistryToken,
+		Insecure:           privateRegistryInsecure,
+		SkipTLS:            privateRegistrySkipTLS,
+		ScenarioRepository: lightspeedRepo,
 	}
 
-	return &response, nil
+	return registry, nil
+}
+
+func NewLightspeedRunCommand(
+	providerFactory *factory.ProviderFactory,
+	scenarioOrchestrator *scenarioorchestrator.ScenarioOrchestrator,
+	config config.Config,
+) *cobra.Command {
+	var command = &cobra.Command{
+		Use:   "run",
+		Short: "Run AI-powered chaos engineering assistance",
+		Long: `Run AI-powered chaos engineering assistance with Retrieval-Augmented Generation (RAG)
+
+This command automatically detects your GPU and deploys a lightweight AI model that can answer 
+questions about krknctl usage, chaos engineering scenarios, and provide intelligent command 
+suggestions based on natural language.
+
+The system uses:
+- Automatic GPU detection (Apple Silicon, NVIDIA)
+- GPU-accelerated inference for fast responses
+- Live documentation indexing
+- Llama 3.2:1B model optimized for chaos engineering domain
+
+Examples:
+  krknctl lightspeed run          # Auto-detect GPU
+  krknctl lightspeed run --no-gpu # Force CPU-only mode`,
+		Args: cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			// Get flags
+			noGPU, _ := cmd.Flags().GetBool("no-gpu")
+
+			// Print container runtime info
+			(*scenarioOrchestrator).PrintContainerRuntime()
+
+			// Check if Docker is being used - Lightspeed only supports Podman
+			if (*scenarioOrchestrator).GetContainerRuntime() == orchestratormodels.Docker {
+				return fmt.Errorf("❌ Lightspeed requires Podman container runtime. Docker is not supported for GPU acceleration")
+			}
+
+			// Get container runtime socket
+			socket, err := (*scenarioOrchestrator).GetContainerRuntimeSocket(nil)
+			if err != nil {
+				return fmt.Errorf("failed to get container runtime socket: %w", err)
+			}
+
+			// Connect to container runtime
+			ctx, err := (*scenarioOrchestrator).Connect(*socket)
+			if err != nil {
+				return fmt.Errorf("failed to connect to container runtime: %w", err)
+			}
+
+			// Build lightspeed registry configuration from flags
+			registry, err := buildLightspeedRegistryFromFlags(cmd, config)
+			if err != nil {
+				return fmt.Errorf("failed to build lightspeed registry configuration: %w", err)
+			}
+
+			// Step 1: Auto-detect GPU acceleration
+			fmt.Println("🔍 Detecting GPU acceleration...")
+
+			// Create platform GPU detector
+			detector := lightspeed.NewPlatformGPUDetector(config)
+			gpuType := detector.DetectGPUAcceleration(ctx, noGPU)
+
+			fmt.Printf("✅ GPU acceleration: %s\n", detector.GetGPUDescription(gpuType))
+
+			// Step 2: Deploy RAG model container
+			fmt.Println("\n🚀 Deploying lightspeed model...")
+
+			// Create spinners for the operations
+			pullSpinner := NewSpinnerWithSuffix(" pulling RAG model image...")
+			thinkingSpinner := NewSpinnerWithSuffix(" thinking...", 35)
+
+			ragResult, err := lightspeed.DeployLightspeedModelWithGPUType(ctx, gpuType, *scenarioOrchestrator, config, registry, detector, pullSpinner)
+			if err != nil {
+				// Handle GPU-related errors with helpful suggestions
+				enhancedErr := detector.HandleContainerError(err, gpuType)
+				return fmt.Errorf("failed to deploy RAG model: %w", enhancedErr)
+			}
+
+			// Step 3: Health check
+			fmt.Println("\n🩺 Performing health check...")
+
+			healthOK, err := lightspeed.PerformLightspeedHealthCheck(ragResult.ContainerID, ragResult.HostPort, *scenarioOrchestrator, ctx, config)
+			if err != nil {
+				return fmt.Errorf("health check failed: %w", err)
+			}
+
+			if !healthOK {
+				return fmt.Errorf("lightspeed service is not responding properly")
+			}
+
+			fmt.Println("✅ Lightspeed service is ready!")
+
+			// Step 4: Start interactive prompt
+			fmt.Printf("\n🤖 Starting interactive Lightspeed service on port %s...\n", ragResult.HostPort)
+			fmt.Println("Type your chaos engineering questions and get intelligent krknctl command suggestions!")
+			fmt.Println("Type 'exit' or 'quit' to stop.")
+
+			return lightspeed.StartInteractivePrompt(ragResult.ContainerID, ragResult.HostPort, *scenarioOrchestrator, ctx, config, thinkingSpinner)
+		},
+	}
+
+	return command
 }
