+++ pkg/lightspeed/gpucheck.go
@@ -1,6 +1,6 @@
 // Package gpucheck provides simple platform-based GPU acceleration detection
 // Generated by Claude Sonnet 4
-package gpucheck
+package lightspeed
 
 import (
 	"context"
@@ -9,20 +9,6 @@ import (
 	"os"
 	"runtime"
 )
-
-// GPUAcceleration represents the type of GPU acceleration available
-type GPUAcceleration string
-
-const (
-	GPUAccelerationAppleSilicon GPUAcceleration = "apple-silicon"
-	GPUAccelerationNVIDIA       GPUAcceleration = "nvidia"
-	GPUAccelerationGeneric      GPUAcceleration = "generic"
-)
-
-// PlatformGPUDetector provides platform-based GPU detection
-type PlatformGPUDetector struct {
-	config config.Config
-}
 
 // NewPlatformGPUDetector creates a new platform-based GPU detector
 func NewPlatformGPUDetector(config config.Config) *PlatformGPUDetector {
@@ -39,11 +25,11 @@ func (d *PlatformGPUDetector) hasNVIDIADevices() bool {
 		"/dev/nvidiactl",
 		"/dev/nvidia-uvm",
 	}
-	
+
 	// At least nvidia0 and nvidiactl should exist for a working NVIDIA setup
 	hasNvidia0 := false
 	hasNvidiaCtl := false
-	
+
 	for _, path := range nvidiaPaths {
 		if _, err := os.Stat(path); err == nil {
 			switch path {
@@ -54,7 +40,7 @@ func (d *PlatformGPUDetector) hasNVIDIADevices() bool {
 			}
 		}
 	}
-	
+
 	// Both essential devices must exist
 	return hasNvidia0 && hasNvidiaCtl
 }
@@ -144,16 +130,16 @@ func (d *PlatformGPUDetector) GetGPUDescription(gpuType GPUAcceleration) string 
 func (d *PlatformGPUDetector) AutoSelectLightspeedConfig(ctx context.Context, noGPU bool) (string, GPUAcceleration, map[string]string, error) {
 	// Detect GPU acceleration type
 	gpuType := d.DetectGPUAcceleration(ctx, noGPU)
-	
+
 	// Get appropriate container image
 	imageURI, err := d.GetLightspeedImageURI(gpuType)
 	if err != nil {
 		return "", "", nil, fmt.Errorf("failed to get lightspeed image for %s: %w", gpuType, err)
 	}
-	
+
 	// Get device mounts
 	deviceMounts := d.GetDeviceMounts(gpuType)
-	
+
 	return imageURI, gpuType, deviceMounts, nil
 }
 
@@ -165,7 +151,7 @@ func (d *PlatformGPUDetector) HandleContainerError(err error, gpuType GPUAcceler
 
 	// Check if this is a GPU-related deployment error
 	errStr := err.Error()
-	
+
 	// Common patterns for GPU/device mounting failures
 	isDeviceError := false
 	if containsAny(errStr, []string{
@@ -193,7 +179,7 @@ func (d *PlatformGPUDetector) HandleContainerError(err error, gpuType GPUAcceler
 	case GPUAccelerationNVIDIA:
 		return fmt.Errorf("%w\n\n💡 NVIDIA GPU acceleration failed. This usually means GPU support is not properly configured.\n\nTo fix this:\n1. Configure GPU support: see https://podman-desktop.io/docs/podman/gpu\n2. Or use CPU-only mode: add --no-gpu flag to your command", err)
 	}
-	
+
 	// Generic device error
 	return fmt.Errorf("%w\n\n💡 GPU device mounting failed. Try using --no-gpu flag for CPU-only mode", err)
 }
